requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.1
torch==2.4.1
pydantic==2.9.2
python-multipart==0.0.12
numpy==1.26.4
pypdf==4.3.1
python-docx==1.1.2
jinja2==3.1.4
